requirements.txt
Flask==0.10.1
PyJWT==1.4.0
Flask-SQLAlchemy==2.4.4
psycopg2>=2.7.5,<2.8.0